pyproject.toml
[project]
name = "drf-tests-ci"
version = "0.1.0"
description = ""
authors = [
    {name = "Anurag Mishra",email = "anurag.mishra@godryft.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-factoryboy = "^2.7.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "sample_project.settings"
python_files = "tests.py test_*.py *_tests.py"